==> script/99_SHUTDOWN.py <==
# %%
import os
import time

# %%
print("shutdown in 10min")
time.sleep(600)
os.system('systemctl poweroff') 
time.sleep(10)
os.system('systemctl poweroff') 
time.sleep(10)
os.system('systemctl poweroff') 
time.sleep(10)
os.system('systemctl poweroff') 
time.sleep(10)
os.system('systemctl poweroff') 
time.sleep(10)
os.system('systemctl poweroff') 
time.sleep(10)
os.system('systemctl poweroff') 
time.sleep(10)
os.system('systemctl poweroff')
